Extract Spotify API client into module
